feat: move getSolutionFromDependance to generator-utils
