+++ packages/generator/generator-utils/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './stripAnsi';
 export * from './package';
 export { getGeneratorPath } from './getGeneratorPath';
+export { getSolutionFromDependance } from './getSolutionFromDependance';
