+++ packages/generator/generator-utils/tests/fixtures/err.json
