+++ packages/generator/generator-utils/tests/fixtures/mwa.json
