+++ packages/generator/generator-utils/tests/fixtures/none.json
