+++ packages/generator/new-action/src/run.ts
@@ -2,11 +2,10 @@
 
 import { Command } from '@modern-js/utils/commander';
 
+import { getSolutionFromDependance } from '@modern-js/generator-utils';
 import { MWANewAction } from './mwa';
 import { ModuleNewAction } from './module';
 import { MonorepoNewAction } from './monorepo';
-
-import { getSolutionByDependance } from './utils';
 
 const main = async () => {
   const program = new Command();
@@ -16,7 +15,7 @@ const main = async () => {
     .option('--root-path <rootPath>', 'project root path', '')
     .action(async params => {
       if (!params.solution) {
-        params.solution = getSolutionByDependance();
+        params.solution = getSolutionFromDependance();
       }
 
       const { solution, config: configStr, rootPath } = params;
