+++ packages/generator/new-action/src/utils/index.ts
@@ -1,5 +1,6 @@
 import path from 'path';
-import { json5, semver } from '@modern-js/utils';
+import { semver } from '@modern-js/utils';
+import { readJsonSync } from '@modern-js/utils/compiled/fs-extra';
 import {
   ActionFunction,
   ActionRefactor,
@@ -7,15 +8,6 @@ import {
   SolutionToolsMap,
 } from '@modern-js/generator-common';
 import { fs, getModernPluginVersion } from '@modern-js/generator-utils';
-
-const swap = (obj: Record<string, string>) => {
-  return Object.keys(obj).reduce<Record<string, string>>((acc, key) => {
-    acc[obj[key]] = key;
-    return acc;
-  }, {});
-};
-
-const dependenceToSolution = swap(SolutionToolsMap);
 
 export function alreadyRepo(cwd = process.cwd()) {
   try {
@@ -25,18 +17,6 @@ export function alreadyRepo(cwd = process.cwd()) {
   }
 }
 
-export const readJson = (jsonPath: string) => {
-  if (!fs.existsSync(jsonPath)) {
-    return {};
-  }
-  const jsonStr = fs.readFileSync(jsonPath, { encoding: 'utf8' });
-  try {
-    return json5.parse(jsonStr);
-  } catch (error) {
-    throw Error(`${jsonPath} is not a valid json, please check and try again.`);
-  }
-};
-
 export function hasEnabledFunction(
   action: ActionFunction | ActionRefactor,
   dependencies: Record<string, string>,
@@ -45,7 +25,7 @@ export function hasEnabledFunction(
   cwd: string,
 ) {
   const packageJsonPath = path.normalize(`${cwd}/package.json`);
-  const packageJson = readJson(packageJsonPath);
+  const packageJson = readJsonSync(packageJsonPath);
   if (!dependencies[action] && !devDependencies[action]) {
     return false;
   }
@@ -84,28 +64,3 @@ export async function usePluginNameExport(
   }
   return true;
 }
-
-export const getSolutionByDependance = (
-  jsonPath?: string,
-): string | undefined => {
-  const packageJsonPath =
-    jsonPath ?? path.normalize(`${process.cwd()}/package.json`);
-  const packageJson = readJson(packageJsonPath);
-
-  const solutions = Object.keys(dependenceToSolution)
-    .map((i: string) => {
-      if (packageJson.dependencies?.[i] || packageJson.devDependencies?.[i]) {
-        return dependenceToSolution[i];
-      }
-      return '';
-    })
-    .filter(Boolean);
-
-  if (solutions.length === 0) {
-    throw new Error('No solution found. Please check your package.json.');
-  } else if (solutions.length > 1) {
-    throw new Error(`Multiple solutions found: ${solutions.join(',')}`);
-  }
-
-  return solutions[0];
-};
